+++ docs/changelog.rst
@@ -23,6 +23,7 @@ Changes in 0.8.X
 - Dynamic fields are now validated on save (#153) (#154)
 - Added support for multiple slices and made slicing chainable. (#170) (#190) (#191)
 - Fixed GridFSProxy __getattr__ behaviour (#196)
+- Fix Django timezone support (#151)
 
 Changes in 0.7.9
 ================